sockitowm:temp28: modify reading temperature function so that it could be used in WRPC, no 1s delays, no floating point

+++ sockitowm/temp28.h
@@ -28,4 +28,4 @@
 //
 // ---------------------------------------------------------------------------
 
-int ReadTemperature28(int,uchar *,float *);
+uint8_t ReadTemperature28(uint8_t , uint8_t *,int16_t *);
